New update

+++ BrainTrainerAPI/Dtos/HighScoreDto.cs
@@ -7,6 +7,6 @@ namespace BrainTrainerAPI.Dtos
         public int UserId { get; set; }
         public int CorrectAnswers { get; set; }
         public int TotalAnswers { get; set; }
-        public IEnumerable<int> TimeBetweenClicksArray { get; set; }
+        public IEnumerable<double> TimeBetweenClicksArray { get; set; }
     }
 }
